TFB/requirements.txt
matplotlib>=3.6.2
numpy>=1.22.3
pandas>=1.5.1
scikit_learn>=1.1.3
scipy>=1.8.0
statsmodels>=0.14.0
ray>=2.6.3
tqdm>=4.64.0
dash>=2.9.3
dash-bootstrap-components>=1.5.0
reformer-pytorch==1.4.4
